requirements.txt
# Browser Automation (from reference repo approach)
selenium>=4.15.0
webdriver-manager>=4.0.0
undetected-chromedriver>=3.5.0

# Telegram Bot
python-telegram-bot>=20.7

# LLM - Groq (primary)
groq>=0.4.0

# LLM - Gemini (fallback)
google-generativeai>=0.3.0

# Config & Utils
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# HTTP
requests>=2.31.0

# Scheduling
schedule>=1.2.0
